backend/requirements.txt
fastapi
uvicorn
scikit-learn
joblib
sqlalchemy
psycopg2-binary
pydantic
python-multipart
requests
python-dotenv